Abstract RowBlockTemplate component Reduce withMaxiProps load Reduce MaxiBlockComponent load Abstracted some MaxiBlockComponent methods Avoid rendering Inspector if it's not selected
