+++ src/extensions/maxi-block/getIsUniqueIDRepeated.js
@@ -5,6 +5,7 @@ import goThroughMaxiBlocks from './goThroughMaxiBlocks';
 
 const getIsUniqueIDRepeated = (uniqueIDToCompare, repeatCount = 1) => {
 	let currentRepeatCount = 0;
+
 	return goThroughMaxiBlocks(block => {
 		if (block.attributes.uniqueID === uniqueIDToCompare) {
 			currentRepeatCount += 1;
