+++ src/extensions/maxi-block/propsObjectCleaner.js
@@ -1,0 +1,26 @@
+// Removes non-necessary entries of props object for comparison
+const propsObjectCleaner = props => {
+	const entriesToRemove = [
+		'maxiSetAttributes',
+		'insertInlineStyles',
+		'cleanInlineStyles',
+		'context',
+		'getBounds',
+	];
+
+	const newProps = Object.entries(props).reduce((acc, [key, value]) => {
+		if (!entriesToRemove.includes(key) && typeof value !== 'function') {
+			if (typeof value === 'object') {
+				acc[key] = JSON.stringify(value);
+			} else {
+				acc[key] = value;
+			}
+		}
+
+		return acc;
+	}, {});
+
+	return newProps;
+};
+
+export default propsObjectCleaner;
