+++ src/extensions/maxi-block/withMaxiProps.js
@@ -26,51 +26,35 @@ const withMaxiProps = createHigherOrderComponent(
 			const { setAttributes, attributes, clientId, isSelected } =
 				ownProps;
 
-			const {
-				deviceType,
-				baseBreakpoint,
-				hasInnerBlocks,
-				isChild,
-				hasSelectedChild,
-				isTyping,
-			} = useSelect(select => {
-				const { receiveMaxiDeviceType, receiveBaseBreakpoint } =
-					select('maxiBlocks');
-				const {
-					getBlockOrder,
-					getBlockParents,
-					hasSelectedInnerBlock,
-					isTyping,
-				} = select('core/block-editor');
+			const { getBlockOrder, getBlockParents } = useSelect(
+				select => select('core/block-editor'),
+				[]
+			);
 
-				const deviceType = receiveMaxiDeviceType();
-				const baseBreakpoint = receiveBaseBreakpoint();
+			const { deviceType, baseBreakpoint, hasSelectedChild } = useSelect(
+				select => {
+					const { receiveMaxiDeviceType, receiveBaseBreakpoint } =
+						select('maxiBlocks');
+					const { hasSelectedInnerBlock } =
+						select('core/block-editor');
 
-				const hasInnerBlocks = !isEmpty(getBlockOrder(clientId));
+					return {
+						deviceType: receiveMaxiDeviceType(),
+						baseBreakpoint: receiveBaseBreakpoint(),
+						hasSelectedChild: hasSelectedInnerBlock(clientId, true),
+					};
+				}
+			);
 
-				const isChild = !isEmpty(
-					getBlockParents(clientId).filter(val => val !== clientId)
-				);
-
-				const hasSelectedChild = hasSelectedInnerBlock(clientId, true);
-
-				return {
-					deviceType,
-					baseBreakpoint,
-					hasInnerBlocks,
-					isChild,
-					hasSelectedChild,
-					isTyping: isTyping(),
-				};
-			});
-
-			const maxiSetAttributes = useCallback(obj =>
-				handleSetAttributes({
-					obj,
-					attributes,
-					clientId,
-					onChange: setAttributes,
-				})
+			const maxiSetAttributes = useCallback(
+				obj =>
+					handleSetAttributes({
+						obj,
+						attributes,
+						clientId,
+						onChange: setAttributes,
+					}),
+				[]
 			);
 
 			const ref = useRef(null);
@@ -90,7 +74,8 @@ const withMaxiProps = createHigherOrderComponent(
 						pseudoElement,
 						styleObjKeys,
 						ref,
-					})
+					}),
+				[]
 			);
 
 			const cleanInlineStyles = useCallback(
@@ -100,7 +85,8 @@ const withMaxiProps = createHigherOrderComponent(
 						pseudoElement,
 						styleObjKeys,
 						ref
-					)
+					),
+				[]
 			);
 
 			const getBounds = useCallback(selector => {
@@ -115,7 +101,7 @@ const withMaxiProps = createHigherOrderComponent(
 				};
 
 				return getTarget().getBoundingClientRect();
-			});
+			}, []);
 
 			useEffect(() => {
 				dispatch('maxiBlocks/styles').savePrevSavedAttrs([]);
@@ -132,20 +118,17 @@ const withMaxiProps = createHigherOrderComponent(
 						getBounds={getBounds}
 						deviceType={deviceType}
 						baseBreakpoint={baseBreakpoint}
-						hasInnerBlocks={hasInnerBlocks}
-						isChild={isChild}
+						hasInnerBlocks={!isEmpty(getBlockOrder(clientId))}
+						isChild={
+							!isEmpty(
+								getBlockParents(clientId).filter(
+									val => val !== clientId
+								)
+							)
+						}
 						hasSelectedChild={hasSelectedChild}
 					/>
-					{/*
-						Need to check if it's typing to avoid an error on Text Maxi when moving the caret selector doing a keyDown event.
-						It happens when, for example, you are typing and you move the caret selector to another block using the arrows.
-					*/}
-					{!isTyping && (
-						<BlockInserter.InterBlockInserter
-							ref={ref}
-							{...ownProps}
-						/>
-					)}
+					<BlockInserter.InterBlockInserter ref={ref} {...ownProps} />
 				</>
 			);
 		}),
